Split code_tree tests into separate file
